new planet

+++ Program.cs
@@ -6,8 +6,8 @@ namespace planetProject
     {
         static void Main(string[] args)
         {
-            double earth_time=31557600;
-            double earth_traveler_time=631154000;
+            long earth_time=31557600L;
+            long earth_traveler_time=631154000;
 
             float earth_in_space_age = earth_traveler_time/earth_time;
 
@@ -17,11 +17,11 @@ namespace planetProject
             planets[2]="Jupiter";
             planets[3]="Neptune";
 
-            double [] ages= new double[4];
-            ages[0]=19414149;
-            ages[1]=59354032;
-            ages[2]=374355659;
-            ages[3]=5200418560;
+            float [] ages= new float[4];
+            ages[0]=19414149F;
+            ages[1]=59354032F;
+            ages[2]=374355659F;
+            ages[3]=5200418560F;
 
 
     for (int i = 0 , j=0; i < planets.Length && j< ages.Length; i++,j++)
